Move the battery prediction calculator into the data model
Relocates utils/battery_prediction.ts to
data/battery_predictions/battery-prediction-calculator.ts so it sits with
the BatteryPredictions singleton it feeds, matching how the rest of the
data model is organized.

Named for what it is rather than as the singular of battery-predictions.ts,
following the data/exclusion_zones/exclusion-zone-detection.ts precedent,
since singular filenames under data/ are entity classes.

Also hoists the endpoint URLs into named constants, drops a non-null
assertion, and corrects a @returns that listed fields the function does
not return.
